requirements.txt
snowflake-connector-python
firecrawl-py
python-dotenv
requests
